Remove the unnecessary cast from pid_t to int

+++ examples/watcher/watcher_main.c
@@ -206,9 +206,9 @@ static int watcher_daemon(int argc, FAR char *argv[])
       goto errout;
     }
 
-  fprintf(fp, "%d %d %d %d %d\n", (int)watcher_pid,
-          (int)CONFIG_EXAMPLES_WATCHER_SIGNAL, (int)SUBSCRIBE_CMD,
-          (int)FEED_CMD, (int)UNSUBSCRIBE_CMD);
+  fprintf(fp, "%d %d %d %d %d\n", watcher_pid,
+          CONFIG_EXAMPLES_WATCHER_SIGNAL, SUBSCRIBE_CMD,
+          FEED_CMD, UNSUBSCRIBE_CMD);
   fclose(fp);
 
   /* Suspends the calling thread until delivery of a non-blocked signal. */
